int殺し。 全ての変更したファイルにわたってintやshortといった型をbit指定のint32_tやint16_tへ変更した。
P.S.見落としで、unsigned char や unsigned shortなどがあるかもしれない。

+++ Matrix4.cpp
@@ -211,11 +211,11 @@ Matrix4::Matrix4(float m00, float m01, float m02, float m03, float m10, float m1
 Matrix4& operator*=(Matrix4& m1, const Matrix4& m2)
 {
     Matrix4 result{ };
-    for (int i = 0; i < 4; i++)
+    for (size_t i = 0; i < 4; i++)
     {
-        for (int j = 0; j < 4; j++)
+        for (size_t j = 0; j < 4; j++)
         {
-            for (int k = 0; k < 4; k++)
+            for (size_t k = 0; k < 4; k++)
             {
                 result.m[i][j] += m1.m[i][k] * m2.m[k][j];
             }
